+++ dx_stream/apps/build.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
+SCRIPT_DIR=$(realpath "$(dirname "$0")")
+PROJECT_ROOT=$(realpath -s "${SCRIPT_DIR}/../..")
 
 BUILD_TYPE="release"
 SONAR_MODE_ARG=""
+NATIVE_FILE_ARG=""
 
 
 show_help() {
@@ -43,13 +46,37 @@ done
 
 echo "Using build type: $BUILD_TYPE"
 
+set_native_file_arg(){
+    local gcc_version=$(gcc -dumpversion | cut -f1 -d.)
+    local gpp_version=$(g++ -dumpversion | cut -f1 -d.)
+
+    echo "Current GCC version: $gcc_version"
+    echo "Current G++ version: $gpp_version"
+
+    # If version is less than 11, install gcc-11 and g++-11
+    if [ "$gcc_version" -lt 11 ] || [ "$gpp_version" -lt 11 ]; then
+        # Check if gcc-11 is installed
+        if dpkg -s gcc-11 >/dev/null 2>&1; then
+            NATIVE_FILE_ARG="--native-file ${PROJECT_ROOT}/gcc11.ini"
+            echo "GCC/G++ version 11 is already installed"
+        else
+            echo "Error: GCC/G++ version 11 or higher are required. Please run install.sh to install them and try again..."
+            exit 1
+        fi
+    else
+        echo "GCC/G++ 11 or higher is already installed."
+    fi
+}
+
 build_and_install() {
+    set_native_file_arg
+
     TARGET_DIR="."
     for subdir in "$TARGET_DIR"/*/; do
         echo "Processing directory: $subdir"
         
         cd "$subdir" || exit 1
-        meson setup build --buildtype="$BUILD_TYPE"
+        meson setup build --buildtype="$BUILD_TYPE" ${NATIVE_FILE_ARG}
         if [ $? -ne 0 ]; then
             echo -e "Error: meson setup failed"
             exit 1
@@ -59,7 +86,7 @@ build_and_install() {
             echo -e "Error: meson compile failed"
             exit 1
         fi
-        sudo meson install -C build
+        yes | meson install -C build
         if [ $? -ne 0 ]; then
             echo -e "Error: meson install failed"
             exit 1