+++ test/pipelines/test_single_pipeline.c
@@ -291,7 +291,7 @@ GST_END_TEST
 Suite *single_suite(void) {
     Suite *s = suite_create("Single TEST");
     TCase *tc_core = tcase_create("Core");
-
+    tcase_set_timeout(tc_core, 20.0);
     tcase_add_test(tc_core, test_single_pipeline);
 
     suite_add_tcase(s, tc_core);